Вживил алгоритм. Работает! Работает!
Доделать
Добавил поля через buddy, проблем нет. Теперь сложность в том, чтобы отображать не id станка например а его модель, чтобы под ним был объект с нужным id.

+++ src/main/java/org/typograf/entity/Work.java
@@ -27,7 +27,7 @@ public class Work {
     private LocalTime timeStart;
 
     @Column(name = "laid_down_time")
-    private Double laidDownTime;
+    private Integer laidDownTime;
 
     public Integer getId() {
         return id;
@@ -69,11 +69,11 @@ public class Work {
         this.timeStart = timeStart;
     }
 
-    public Double getLaidDownTime() {
+    public Integer getLaidDownTime() {
         return laidDownTime;
     }
 
-    public void setLaidDownTime(Double laidDownTime) {
+    public void setLaidDownTime(Integer laidDownTime) {
         this.laidDownTime = laidDownTime;
     }
